Refactor WPitchInterval.inverse_direction->inverse
This commits refactors the method
'mutwo.music_parameters.WesternPitchInterval.inverse_direction' to
'mutwo.music_parameters.WesternPitchInterval.inverse'.

Why?
====

Interval inversion in musicology is actually the switching of interval
direction. And the much older mutwo method "JustIntonationPitch.inverse"
already does exactly this. So before this commit we had two different method
names for the same behaviour:

  my_just_intonation_pitch.inverse().interval == - my_just_intonation_pitch.interval
  my_western_pitch_interval.inverse_direction().interval == - my_western_pitch_interval.interval

But now it is also:

  my_western_pitch_interval.inverse().interval == - my_western_pitch_interval.interval

which makes the whole library more consistent because now 'inverse' for
JustIntonationPitch has exactly same meaning like 'inverse' for
WesternPitchInterval.

In music21 the method with this semantic is called "reverse":

    https://web.mit.edu/music21/doc/moduleReference/moduleInterval.html#music21.interval.Interval.reverse

But music21 also knows "mod7inversion":

    https://web.mit.edu/music21/doc/moduleReference/moduleInterval.html#music21.interval.DiatonicInterval.mod7inversion

The semantics of "mod7inversion" is currently not yet supported by
mutwo. This is the reason why initially the WesternPitchInterval method
was called "inverse_direction", so "inverse" would still be free for the
sematnics of "mod7inversion". But "mod7inversion" is actually a much
more specific type of inversion. It should be implemented on a different
name than "inverse".