music_parameters: Add musical scales
Starting from now it is possible to represent musical scales in mutwo
using the mutwo.music package.

The scale implementation is very similar to the one used in music21:

We have a "ScaleFamily" which is mostly the same like music21's
"AbstractScale" -> So it only contains intervals without any specific
tonic.

Then we have "Scale" which is mostly the same like music21's
"ConcreteScale". It's simply a mixture between a scale family which has
been assigned to one specific pitch. So a scale also doesn't only
contain intervals, but actual pitches.