JustIntoPitch: Make comparable with PitchInterval
Before this patch JustIntonationPitch couldn't be compared with any
other PitchInterval object.

The origin for this problem is that a JustIntonationPitch is both, a
PitchInterval and a Pitch and when comparing it always uses the pitch
comparison methods. It should actually use the PitchInterval comparison
methods if it is set in relation to another PitchInterval instance. The
real fix for this problem would be to fix `mutwo.core` to support
`SingleValueParameter` to allow multi-inheritance. But this fix can't be
done without an API change. Therefore we need to wait until
mutwo.core v.2.0.0 release.

This patch fixes issue https://github.com/mutwo-org/mutwo.music/issues/2