parameters: Add Pitch->Fingering func; add .sound
One reason to specify fingerings is to map certain chords to
certain fingerings. The new method pitch_sequence_to_fingering_tuple
helps here. It's empty by default (abc method would lead to a lot
of boilerplate in builtin instruments until function is properly
implemented there, which takes a lot of time).

We need a new 'Fingering.Part.sound' attribute in order to assign
pitches to fingering parts. This is needed because when creating
fingerings for chords some fingerings may include fingers of both
hands. Because left/right hand are usually notated in different
staves in piano/harp music we need to know which fingering part
(-> hand) maps to which sound in order to split the pitches later
to the responsible staff.