instruments: Refactor + make ortools optional
Currently it's not possible to call `nix-build` on a package which
depends on ortools. See:

    https://discourse.nixos.org/t/ortools-module-not-found-under-nix-shell/14190/7
and

    https://stackoverflow.com/questions/68477623/ortools-not-found-in-nix-shell

When installing a package with 'python310Packages.ortools' in 'propagatedBuildInputs',
setuptools complains that it doesn't find ortools.

---

In order to make 'mutwo.music' installable, we make 'ortools' package
optional.

This is generally a good idea, because it's a rather difficult to
install package (needs to be compiled, non-python source code, etc.)
and we only use it for one very specific method.

---

In order to make it optional, it's better to refactor instruments to
make the optional block better readable (it's all in
mutwo/music_parameters/instruments/CelticHarp.py now).