Add PitchedInstrument.get_pitch_variant_tuple
It's better to define an explicit 'get_pitch_variant_tuple' for
pitched instruments. We can't rely on
'instrument.pitch_ambitus.get_pitch_variant_tuple', because for
a 'DiscreetPitchedInstrument' this could return pitches which are
actually impossible to be played on the given instrument. For instance
a 'cqs4' is in the ambitus of a piano and we can imagine many variants
of it within its ambitus, but it is nevertheless unplayable, since
piano can't play quartertones (ok, most of the time they can't).