Replaced deprecated natural harmonic indicators
Before this patch we had a indicator for a natural harmonic, which
actually only changed the note head and a second class which changed the
note head and wrote the string pitch. Both indicators weren't capable of
writing double harmonics.

In order to simplify and unify the code, both have been removed and
replaced by `NaturalHarmonicList`, which can represent all cases which
were represented before and even more.