parameters: Add .natural_harmonic to Node
It's useful if we know the natural harmonic from which a node is
created, because than we can not only figure out where to press a
string, but we also know which pitch is produced by pressing the finger
at this position. It's handy to know both of these properties within one
object, because otherwise we would always need to carry both a node and
a natural harmonic to the converters, so that they could represent both
in the final score.