generators: Drop wilson module
Functions related to the tuning theorist Erv Wilson partially depend on
the 'mutwo.common' module. 'mutwo.music' has therefore only because of
this an optional dependency on 'mutwo.common'. It's better to have
simpler, smaller packages for easier maintenance, so we should drop the
'generators.wilson' module from 'mutwo.music'. If anyone wants to use
the given code, this person can do so by adding a custom 'mutwo.wilson'
package.