converters: Drop pitches submodule
This submodule contained a converter which helped re-writing accidentals of
chromatic music in a way that the music is easier to read (e.g. that the
accidentals look as tonal as possible). While such a converter is
generally a good idea and neither should be too complicated to
implement, the current implementation had two big issues:

    (1) It was dependent on a non-trivial dependency package.
    (2) It didn't work that well.

Because I currently don't have time to rework this (and I don't need it
for any current project), the easiest way is to simply drop this
converter entirely. In this way 'mutwo.music' becomes easier to install
and easier to maintain on a longer term. Of course, if this converter
would be needed by anyone (including myself), a merge request could be
created and potentially some bits from the now-dropped could be useful.
But until then, let's drop this.