events/NoteLike: Add syntactic sugar for indicator collections
Many arguments of 'NoteLike' support a variety of types: for instance
'pitch_list' supports string, lists, tuplets, pitch objects, numbers
(float or fraction) and more. In this way the 'NoteLike' usage can be
more convenient. For indicator collections (playing and notation
indicators) such syntactic sugar was missing before this patch, which
made the assignments of indicator collection very cumbersome. This patch
improves the situation by adding a parser from strings to indicator
collections. This doesn't only simplify the general usage of 'NoteLike',
but is particularly useful when parsing text files into mutwo structure,
as it's done by 'mutwo.mmml'.

This is done by reading the input text by an adapted version of
builtin configparser. In this way we don't need to quote strings
(as it'd be necessary for json or toml parser and therefore make
the usage less concise again) and we also don't need yaml (for which
python stdlib doesn't provide a parser, which means we'd need one more
dependency again).