Drop converters.LoudnessToAmplitude + related bits
The converter from loudness to amplitude may be useful, but it depends
on two heavy packages: numpy and scipy. Furthermore it could be argued,
that it doesn't exactly belong to music itself, but more to something
like 'acoustic'. A lot of people who may want to use the music
representations, may very likely not need the functionality to convert
from loudness to amplitude. I personally only needed it for 1 from 10
projects. The combination of this high speciality and the necessity of
heavy dependencies lead to the decision to drop this converter from
'mutwo.music' and to move this functionality to the new package
'mutwo.acoustic'. As this was the last code that needed 'numpy' and
'scipy' in 'mutwo.music', we can now also drop those dependencies from
'mutwo.music' and therefore make the package lighter.