Update to mutwo.core >= 2.0.0: Let tests pass
This is the first step in updating 'mutwo.music' to the new version of
'mutwo.core'. This patch quite aggressively simply adjust the source
code so that all tests of 'mutwo.music' pass with the new 'mutwo.core'
version. In order to do this the glissando functionality of pitches
needed to be removed. In the next steps, we need to re-introduce this
glissando functionality and we also need to apply step-by-step
'mutwo.music' parts of the refactoring procedures that happened from
'mutwo.core' == 1.X.Y to 2.0.0.