Update to mutwo.core >= 2.0.0: Refactor syntactic sugar parser
This is the 'mutwo.music' equivalent to https://github.com/mutwo-org/mutwo.core/commit/35e05cd3722d4d65d7593c047d2918372cf301e6.
It changes all 'UNKNOWN_OBJECT_TO_X' calls to
'music_parameters.abc.X.from_any'.

NOTE: With this patch we broke the possibility to add custom playing and
notation indicators and a custom Playing- or
NotationIndicatorCollection that when set to
DEFAULT_X_INDICATOR_COLLECTION is used by 'NoteLike', because in the
indicator_collection setters 'NoteLike' no longer relates to the
'DEFAULT_X_INDICATOR_COLLECTION'. A fix for this would be to add a new
registry functionality for indicators: in this case each
'IndicatorCollection' would host a registry and each indicator would be
registered in-place, instead of the collection that explicitly assigns
them to itself.

This patch furthermore removes two previously existing functionalities:

1. Indicator collections don't behave as if frozen anymore: setting
   parameters no longer raises a FrozenError anymore. This should be ok,
   as soon we should rewrite the logic of indicator collections anyway.

2. It's no longer possible to set Chronon to (after_)grace_note_consecution
   parameter of NoteLike.