parameters/IndicatorCollection: Add registry logic
In projects, we often need to add custom playing or notation indicators,
that are too specific to be included in 'mutwo.music'. Until https://github.com/mutwo-org/mutwo.music/commit/500387bcac9db09f025f9e55b068970af5b9e44d
it was possible to define a new 'PlayingIndicatorCollection' or
'NotationIndicatorCollection' with the new indicator and to override
the 'DEFAULT_X_INDICATOR_COLLECTION'. In this way, 'NoteLike' could be
used with the new custom indicator.

Since https://github.com/mutwo-org/mutwo.music/commit/500387bcac9db09f025f9e55b068970af5b9e44d this is no longer possible.
It was dropped because it conflicted with the new design of syntactic
sugar parsers. But it was also dropped, because it wasn't worth it to
keep it longer, because this API was quite cumbersome and unclear (implicit).

This patch adds a new (explicit) API to add new indicator types, which is
much easier to use. It adds the new '.register' class method to
'IndicatorCollection'. Using this '.register' function, new indicator
types can be added to collections. When adding a new indicator in a
project, we therefore now only need to add '.register' as a decorator
before the indicator definition:

>>> @music_parameters.PlayingIndicatorCollection.register
... @dataclasses.dataclass
... class Noise(music_parameters.abc.ImplicitPlayingIndicator):
...     pitch_name: typing.Optional[str] = "c4"
...     is_active: bool = False