parameters/Volume.from_any: allow parsing strings not only for WesternVolume
Before this patch, it was only possible to create 'WesternVolume'
objects, when providing a string value to the volume parser.
After this patch, the volume parser is more clever and it tries to
evaluate everything that doesn't look like a 'WesternVolume' definition
into a normal python object, in order to follow the standard volume
parsing logic. This is useful for 'mutwo.mmml'.